+++ docs/api/cdktf.md
@@ -720,12 +720,12 @@ When given a project, this it the project itself.
 | <code><a href="#projen.cdktf.ConstructLibraryCdktf.property.libdir">libdir</a></code> | <code>string</code> | The directory in which compiled .js files reside. |
 | <code><a href="#projen.cdktf.ConstructLibraryCdktf.property.srcdir">srcdir</a></code> | <code>string</code> | The directory in which the .ts sources reside. |
 | <code><a href="#projen.cdktf.ConstructLibraryCdktf.property.testdir">testdir</a></code> | <code>string</code> | The directory in which tests reside. |
-| <code><a href="#projen.cdktf.ConstructLibraryCdktf.property.tsconfigDev">tsconfigDev</a></code> | <code>projen.javascript.TypescriptConfig</code> | A typescript configuration file which covers all files (sources, tests, projen). |
+| <code><a href="#projen.cdktf.ConstructLibraryCdktf.property.tsconfigDev">tsconfigDev</a></code> | <code>projen.typescript.TypescriptConfig</code> | A typescript configuration file which covers all files (sources, tests, projen). |
 | <code><a href="#projen.cdktf.ConstructLibraryCdktf.property.watchTask">watchTask</a></code> | <code>projen.Task</code> | The "watch" task. |
 | <code><a href="#projen.cdktf.ConstructLibraryCdktf.property.docgen">docgen</a></code> | <code>boolean</code> | *No description.* |
 | <code><a href="#projen.cdktf.ConstructLibraryCdktf.property.eslint">eslint</a></code> | <code>projen.javascript.Eslint</code> | *No description.* |
-| <code><a href="#projen.cdktf.ConstructLibraryCdktf.property.tsconfig">tsconfig</a></code> | <code>projen.javascript.TypescriptConfig</code> | *No description.* |
-| <code><a href="#projen.cdktf.ConstructLibraryCdktf.property.tsconfigEslint">tsconfigEslint</a></code> | <code>projen.javascript.TypescriptConfig</code> | *No description.* |
+| <code><a href="#projen.cdktf.ConstructLibraryCdktf.property.tsconfig">tsconfig</a></code> | <code>projen.typescript.TypescriptConfig</code> | *No description.* |
+| <code><a href="#projen.cdktf.ConstructLibraryCdktf.property.tsconfigEslint">tsconfigEslint</a></code> | <code>projen.typescript.TypescriptConfig</code> | *No description.* |
 
 ---
 
@@ -1421,7 +1421,7 @@ The directory in which tests reside.
 public readonly tsconfigDev: TypescriptConfig;
 ```
 
-- *Type:* projen.javascript.TypescriptConfig
+- *Type:* projen.typescript.TypescriptConfig
 
 A typescript configuration file which covers all files (sources, tests, projen).
 
@@ -1465,7 +1465,7 @@ public readonly eslint: Eslint;
 public readonly tsconfig: TypescriptConfig;
 ```
 
-- *Type:* projen.javascript.TypescriptConfig
+- *Type:* projen.typescript.TypescriptConfig
 
 ---
 
@@ -1475,7 +1475,7 @@ public readonly tsconfig: TypescriptConfig;
 public readonly tsconfigEslint: TypescriptConfig;
 ```
 
-- *Type:* projen.javascript.TypescriptConfig
+- *Type:* projen.typescript.TypescriptConfig
 
 ---
 
@@ -1672,8 +1672,8 @@ const constructLibraryCdktfOptions: cdktf.ConstructLibraryCdktfOptions = { ... }
 | <code><a href="#projen.cdktf.ConstructLibraryCdktfOptions.property.sampleCode">sampleCode</a></code> | <code>boolean</code> | Generate one-time sample in `src/` and `test/` if there are no files there. |
 | <code><a href="#projen.cdktf.ConstructLibraryCdktfOptions.property.srcdir">srcdir</a></code> | <code>string</code> | Typescript sources directory. |
 | <code><a href="#projen.cdktf.ConstructLibraryCdktfOptions.property.testdir">testdir</a></code> | <code>string</code> | Jest tests directory. Tests files should be named `xxx.test.ts`. |
-| <code><a href="#projen.cdktf.ConstructLibraryCdktfOptions.property.tsconfig">tsconfig</a></code> | <code>projen.javascript.TypescriptConfigOptions</code> | Custom TSConfig. |
-| <code><a href="#projen.cdktf.ConstructLibraryCdktfOptions.property.tsconfigDev">tsconfigDev</a></code> | <code>projen.javascript.TypescriptConfigOptions</code> | Custom tsconfig options for the development tsconfig.json file (used for testing). |
+| <code><a href="#projen.cdktf.ConstructLibraryCdktfOptions.property.tsconfig">tsconfig</a></code> | <code>projen.typescript.TypescriptConfigOptions</code> | Custom TSConfig. |
+| <code><a href="#projen.cdktf.ConstructLibraryCdktfOptions.property.tsconfigDev">tsconfigDev</a></code> | <code>projen.typescript.TypescriptConfigOptions</code> | Custom tsconfig options for the development tsconfig.json file (used for testing). |
 | <code><a href="#projen.cdktf.ConstructLibraryCdktfOptions.property.tsconfigDevFile">tsconfigDevFile</a></code> | <code>string</code> | The name of the development tsconfig.json file. |
 | <code><a href="#projen.cdktf.ConstructLibraryCdktfOptions.property.tsJestOptions">tsJestOptions</a></code> | <code>projen.typescript.TsJestOptions</code> | Options for ts-jest. |
 | <code><a href="#projen.cdktf.ConstructLibraryCdktfOptions.property.typescriptVersion">typescriptVersion</a></code> | <code>string</code> | TypeScript version to use. |
@@ -3772,7 +3772,7 @@ compile the code in-memory.
 public readonly tsconfig: TypescriptConfigOptions;
 ```
 
-- *Type:* projen.javascript.TypescriptConfigOptions
+- *Type:* projen.typescript.TypescriptConfigOptions
 - *Default:* default options
 
 Custom TSConfig.
@@ -3785,7 +3785,7 @@ Custom TSConfig.
 public readonly tsconfigDev: TypescriptConfigOptions;
 ```
 
-- *Type:* projen.javascript.TypescriptConfigOptions
+- *Type:* projen.typescript.TypescriptConfigOptions
 - *Default:* use the production tsconfig options
 
 Custom tsconfig options for the development tsconfig.json file (used for testing).
