+++ src/typescript/index.ts
@@ -1,5 +1,5 @@
 export * from "./projenrc";
 export * from "./projenrc-ts";
-export * from "./typescript-config";
+export * from "../javascript/typescript-config";
 export * from "./typescript";
 export * from "./typescript-typedoc";
